ref(hooks): Move `useTheme` hook to `/lib/hooks/`
