+++ src/layout/Footer.tsx
@@ -57,7 +57,7 @@ export default function Footer() {
           <ul className="space-y-2 text-sm">
             <li><a href="/#about" className={underlineCls}>Tentang Sekolah</a></li>
             <li><a href="/a/program-keahlian" className={underlineCls}>Program Keahlian</a></li>
-            <li><a href="/a/pelatihan-vokasi-2025" className={underlineCls}>Pelatihan Vokasi 2025</a></li>
+            <li><a href="/a/analisis-vokasi-2025" className={underlineCls}>Analisis Vokasi 2025</a></li>
             <li><a href={mainWebUrl} target="_blank" rel="noopener noreferrer" className={underlineCls}>PPDB</a></li>
           </ul>
         </div>
